### 2017-12-06 #488 @whedon
[REVIEW]: CheSweet: An application to predict glycan’s chemicals shifts
**Submitting author:** @PabloGGaray (<a href="http://orcid.org/0000-0001-6604-0329">Pablo Garay</a>)
**Repository:** <a href="https://github.com/BIOS-IMASL/chesweet" target ="_blank">https://github.com/BIOS-IMASL/chesweet</a>
**Version:** 0.0.1
