### 2017-12-06 #482 @arfon
@richardjgowers - please head over to #488 to carry out the actual review. I'll be following along there so just shout if anything isn't clear.
