### 2017-12-07 #488 @PabloGGaray
Hello @arfon, excuse me if this is not the place to do this question, but I have to add two line to the paper with the funding of the project, how I do that? I wait for the review or I can add that lines now?.
Thanks in advance and excuse me for the inconveniences.
