### 2017-12-07 #488 @arfon
@PabloGGaray - authors do this pretty regularly. I would suggest adding a new section header to the `paper.md` file:

```
# Funding

Anything you want to say about the funding source for the project...
```
