### 2017-12-08 #488 @PabloGGaray
I made the change, I suppose you have to compile the `paper.md` again for the review, thanks.
