### 2017-12-13 BIOS-IMASL/chesweet#15 @richardjgowers
Where to seek support
This is related to the JOSS review at openjournals/joss-reviews#488

It's not clear if you can raise an issue here for support or if there's an alternate place for support (eg board/email etc)
