### 2017-12-13 BIOS-IMASL/chesweet#16 @richardjgowers
Docs for running tests
This is related to the JOSS review at openjournals/joss-reviews#488

The installation instructions are designed for cloning the repo but then not cd'ing into it.  But the instructions for pytest assume you are inside repo (running outside of the repo fails as there is a relative path used for lut).  This could be clarified
