### 2017-12-13 BIOS-IMASL/chesweet#18 @richardjgowers
How to write lookup table
This is related to the JOSS review at openjournals/joss-reviews#488

Re this point:
Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I couldn't find documentation for how to write lookup tables yourself, ie what's the format required for these files

Slightly related, the tables included would benefit from a header line explaining what the columns all are, rather than having to consult the code to see how they're read/used to figure this out
