### 2017-12-13 #488 @richardjgowers
@arfon I've raised some issues on their repo, overall the paper + software looks good - I'd consider this minor revisions.

I also need to double check the maths of what's happening (functional claims of software, ie are the interpolations all behaving as they claim), but I can do this while they address the issues I've raised.
