### 2017-12-15 #488 @arfon
Great, thanks for the update @richardjgowers.

> I also need to double check the maths of what's happening (functional claims of software, ie are the interpolations all behaving as they claim), but I can do this while they address the issues I've raised.

👍 let me know when you're done and satisfied with the submission.
