### 2017-12-15 ba47247
added Suport information and information about lut files
