+++ README.md
@@ -62,7 +62,7 @@ import chesweet as chsw
 ```
 
 ### Load look-up tables (lut)
-One of the main argument of `load` is `full`. Passing `False` is useful when working with the glycosidic torsional angles Φ, Ψ and Ω, `True` allows you to work with the mentioned torsional angles and also included the *Χ* torsionals. In the image below you can see the mentioned torsional angles for α-D-Glcp-(1-4)-α-D-Glcp.  
+One of the main argument of `load` is `full`. Passing `False` (default) is useful when working with the glycosidic torsional angles phi (Φ), psi (Ψ), and omega (Ω), `True` allows you to work with the mentioned torsional angles and also included the chi (*Χ*) torsionals. In the image below you can see the mentioned torsional angles for α-D-Glcp-(1-4)-α-D-Glcp.  
 
 <p align="center">
 <img src="img/maltose_all_torsional.png", width=500>
@@ -73,7 +73,7 @@ The names of the disaccharides used by *Che*Sweet follow the IUPAC syntax, but w
 Examples:  
 Load **all calculated dissacharides** in *Che*Sweet:  
 ```python
-# Reduced version of the datasets (using only phi$, psi and omega)
+# Reduced version of the datasets (using only phi, psi and omega)
 disaccharides_red = chsw.CheSweet()
 
 # Using full=True (using all torsionals)
@@ -100,8 +100,67 @@ disaccharides_red = chsw.CheSweet(disaccharides = disaccharides_list)
 disaccharides_full = CheSweet(disaccharides = disaccharides_list, full=True)
 ```
 
-Once you have loaded the lool-up tables it is possible to calculate the chemical shifts or the torsional that you want.  
-In the next examples we show how to use these functions for maltose \[α-D-Glcp-(1-4)-α-D-Glcp]
+#### Format of the look-up table files
+
+For each look-up table file (lut file, for short) the last two columns are the values of the pre-calculated CS. The first of that columns corresponds to the C1 and the last correspond to the second carbon in the glycosidic bond (Cx). The remaining columns are the torsional angles.  
+
+There are two types of lut files for each disaccharide, one type (called *"full"*) includes all the torsional angles (lut files with the name of the disaccharide) and the other type (called *"reduce"*) only includes the torsionals of the glycosidic bond (lut files with the name ended with "_red"). Which of this files is used in the calculation is controlled with the keyword `full`. 
+
+Last 4 lines in the file "a-D-Glcp-1-1-a-D-Glcp", the header with the columns names is not present in the actual lut files and is show here only to help identify the content of each column:  
+```
+  phi psi  chi1 chi2 nan  nan    C1 CS    C1' CS
+ -10  130  -60  -60    0    0    87.3016  87.6553
+ -10  140   60   60    0    0    85.9775  86.1810
+ -10  140  180   60    0    0    86.0382  85.3786
+ -10  140  -60   60    0    0    84.7857  85.0489
+```
+
+Here the chi1 belongs to the first residue and chi2 to the second residue
+
+Last 4 lines in the file "a-D-Glcp-1-4-a-D-Glcp":  
+```
+  phi psi  chi1 chi2  chi3  nan  C1 CS    C4' CS
+ -10  130  180  -60  -60    0    89.1629 117.2891
+ -10  130  -60   60  -60    0    88.2327 117.6398
+ -10  130  -60  180  -60    0    88.2642 115.4036
+ -10  130  -60  -60  -60    0    87.5465 116.9395
+```
+
+Here chi1 belongs to the first residue and chi2-chi3 to the second residue
+
+Last 4 lines in the file "a-D-Glcp-1-6-a-D-Glcp":  
+```
[…]
