### 2017-12-20 ac4b9c7
two section of README reordered (#20)
