### 2017-12-21 fb20c93
changes in the way of how the package and is files are loaded
