+++ chesweet/test/test_chesweet.py
@@ -1,4 +1,3 @@
-#from numpy.testing import assert_almost_equal
 import numpy as np
 from ..chesweet import *
 from ..chesweet import _round_down_up, _nearest_chi
@@ -8,9 +7,9 @@ torsionals = [-10, 10., 180., -180., 60., 75., -185, 185]
 ef_corr = 183.4
 disaccharides_list = ['a-D-Glcp-1-1-a-D-Glcp', 'a-D-Galp-1-3-b-D-Galp', 'b-D-Galp-1-6-b-D-Galp']
 
-disaccharides_red = CheSweet(path='chesweet/lut', disaccharides = disaccharides_list)
+disaccharides_red = CheSweet()
 
-disaccharides_full = CheSweet(path='chesweet/lut', disaccharides = disaccharides_list, full=True)
+disaccharides_full = CheSweet(full=True)
 
 ### Data to test compute_cs function
 
@@ -43,8 +42,8 @@ full_chem_1_6 = [(72.9099, 113.8796), (76.7441, 112.9245), (np.inf, np.inf),
                  (77.953199, 109.076416), (72.6276, 112.8706), 
                  (np.inf, np.inf)]
 
-red_tors_1_6 = [(-70, 120, 180), (135, 65, 60), (0, -10, 60), 
-                 (-126.4, 46.7, 46.8), (-155.7, -76.4, 160.8), (-175.89, -26.7, 78.6)]
+red_tors_1_6 = [(-70, 120, 180), (135, 65, 60), (0, -10, 60), (-126.4, 46.7, 46.8),
+                (-155.7, -76.4, 160.8), (-175.89, -26.7, 78.6)]
 red_chem_1_6 = [(72.4989, 113.5847),(77.6637, 113.626), (np.inf, np.inf), 
                 (79.505894, 109.841949), (71.4614, 112.203), (np.inf, np.inf)]
 
@@ -74,8 +73,7 @@ tors_test_full_1_3 = [np.array([[ 100., -120.,  -60.,  -60.,   60.],
                                 [ 140., -120.,  180.,  180.,  -60.]]),
                       np.zeros(shape=(0, 5)),
                       np.zeros(shape=(0, 5)),
-                      np.zeros(shape=(0, 5))
-                     ]# np.zeros only used to completed the list
+                      np.zeros(shape=(0, 5))]  # np.zeros only used to completed the list
 
 chem_test_full_1_1 = [(82.4441, 90.0294), (20, 150), (82.4441, 150),
                       (20, 90.0294)]
@@ -93,8 +91,7 @@ tors_test_full_1_1 = [np.array([[-120.,  100.,   60.,  180.],
                                 [ -70.,   90.,  180.,   60.]]),
                       np.zeros(shape=(0, 5)),
                       np.zeros(shape=(0, 5)),
-                      np.zeros(shape=(0, 5))
-                     ]
+                      np.zeros(shape=(0, 5))]
 
 chem_test_full_1_6 = [(68.7441, 112.9245), (20, 150), (68.7441, 150),
                       (20, 112.9245)]
@@ -107,8 +104,7 @@ tors_test_full_1_6 = [np.array([[  40.,  -20.,   60.,  180.,  -60.],
                                 [ -80., -120.,  180.,  180.,  -60.]]),
                       np.zeros(shape=(0, 5)),
                       np.zeros(shape=(0, 5)),
-                      np.zeros(shape=(0, 5))
-                     ]
+                      np.zeros(shape=(0, 5))]
 
 chem_test_red_1_3 = [(71.9521, 91.3869), (20, 150), (71.9521, 150),
                      (20, 91.3869)]
@@ -118,8 +114,7 @@ tors_test_red_1_3 = [np.array([[ 110., -120.],
[…]
