### 2017-12-21 68d1d72
Changes in the way of how the package and his files are loaded (#21)
* changes in the way of how the package and is files are loaded

* little change in tests

* undo previous commit
