### 2018-01-12 BIOS-IMASL/chesweet#18 @PabloGGaray
Hi @richardjgowers, we made the changes that you suggested in this Issue and in the Issues #15, #16 and #17. Let us know if these changes are according to your suggestions.
