### 2018-01-22 BIOS-IMASL/chesweet#16 @richardjgowers
This was fixed by #17
