### 2018-01-25 #488 @arfonsmith
Sure. Please make the edits to the paper.md file and I can regenerate the
PDF.

On Thu, Jan 25, 2018 at 1:04 PM, Pablo G Garay <notifications@github.com>
wrote:

> Hello @arfon <https://github.com/arfon>, Its possible make little changes
> in the paper?
> We committed a little mistake in the affiliation information, the name of
> our institution is "Instituto de Matemática *Aplicada* San Luis", we
> omitted the "Aplicada" in the name by mistake.
> And, if it is possible another change, in one reference the names of the
> authors don't appear, but we don't know the cause of this last problem.
>
