### 2018-01-27 #488 @PabloGGaray
We made the change @arfon, the paper.md is ready for the regeneration of the PDF.
