### 2018-01-27 #488 @arfon
Great. I've updated the paper now. It may take a few hours to show the updated version on the JOSS site as the papers are cached.
